merger-tracker/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
fastapi-cache2[memory]==0.2.1
resend==0.8.0
apscheduler==3.10.4
pydantic-settings==2.1.0
